Revert "Русские и усиленные Стражи v.2.0." (#2411)
Revert "Tweak: adjust guardians stats (#2330)"

This reverts commit c83cfa2be25ce735d808da36d52d1f1339b07257.

+++ code/game/gamemodes/miniantags/guardian/types/lightning.dm
@@ -30,17 +30,18 @@
 
 /mob/living/simple_animal/hostile/guardian/beam/AttackingTarget()
 	. = ..()
-	if(. && isliving(target) && target != src && target != summoner)
-		cleardeletedchains()
-		for(var/chain in enemychains)
-			var/datum/beam/B = chain
-			if(B.target == target)
-				return //oh this guy already HAS a chain, let's not chain again
-		if(enemychains.len > 2)
-			var/datum/beam/C = pick(enemychains)
-			qdel(C)
-			enemychains -= C
-		enemychains += Beam(target, "lightning[rand(1,12)]", 'icons/effects/effects.dmi', time=70, maxdistance=7, beam_type=/obj/effect/ebeam/chain)
+	if(prob(10))
+		if(. && isliving(target) && target != src && target != summoner)
+			cleardeletedchains()
+			for(var/chain in enemychains)
+				var/datum/beam/B = chain
+				if(B.target == target)
+					return //oh this guy already HAS a chain, let's not chain again
+			if(enemychains.len > 2)
+				var/datum/beam/C = pick(enemychains)
+				qdel(C)
+				enemychains -= C
+			enemychains += Beam(target, "lightning[rand(1,12)]", 'icons/effects/effects.dmi', time=20, maxdistance=4, beam_type=/obj/effect/ebeam/chain)
 
 /mob/living/simple_animal/hostile/guardian/beam/Destroy()
 	removechains()
@@ -114,7 +115,7 @@
 						"<span class='userdanger'>You are shocked by the lightning chain!</span>", \
 						"<span class='italics'>You hear a heavy electrical crack.</span>" \
 					)
-				L.adjustFireLoss(4)
+				L.adjustFireLoss(3)
 				. = 1
 
 /mob/living/simple_animal/hostile/guardian/beam/death(gibbed)
